refactor(state-migration): separate system-text-json implementation
