+++ Egil.Orleans.StateMigration/test/Egil.Orleans.StateMigration.Tests/Egil.Orleans.StateMigration.Tests.csproj
@@ -14,6 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Using Include="Egil.Orleans.StateMigration.SystemTextJson" />
     <Using Include="Xunit" />
   </ItemGroup>
