+++ Egil.Orleans.StateMigration/test/Egil.Orleans.StateMigration.Tests/SystemTextJson/OrleansInProcessClusterFixture.cs
@@ -1,7 +1,7 @@
 using System.Text.Json;
 using Orleans.TestingHost;
 
-namespace Egil.Orleans.StateMigration.Tests;
+namespace Egil.Orleans.StateMigration.Tests.SystemTextJson;
 
 public sealed class OrleansInProcessClusterFixture : IAsyncLifetime
 {
