+++ Egil.Orleans.StateMigration/test/Egil.Orleans.StateMigration.Tests/SystemTextJson/StateMigrationJsonSerializerOptionsExtensionsTests.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
 
-namespace Egil.Orleans.StateMigration.Tests;
+namespace Egil.Orleans.StateMigration.Tests.SystemTextJson;
 
 public sealed class StateMigrationJsonSerializerOptionsExtensionsTests
 {
